requirements.txt
imageio==2.19.3
loguru==0.6.0
matplotlib==3.5.2
numpy==1.22.4
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
torch==1.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.64.0
transforms3d==0.3.1
